django-schools-master/django_school/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.9.1
black==19.10b0
click==7.1.2
Django==3.0.8
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==14.1.0
django-bootstrap4==2.2.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-js-asset==1.2.2
django-phonenumber-field==4.0.0
pathspec==0.8.0
phonenumbers==8.12.6
Pillow==7.2.0
pytz==2020.1
regex==2020.6.8
selenium==3.141.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
